refactor(components): replace and remove old button components
