--- src/components/Buttons/index.ts
@@ -1,3 +1,0 @@
-export { default as Button, type ButtonProps, checkIsSecondary } from "./Button"
-export { default as ButtonLink, type ButtonLinkProps } from "./ButtonLink"
-export { default as IconButton } from "./IconButton"
