+++ src/components/Staking/StakingHierarchy.tsx
@@ -7,7 +7,6 @@ import { ChildOnlyProp } from "@/lib/types"
 import { cn } from "@/lib/utils/cn"
 import { trackCustomEvent } from "@/lib/utils/matomo"
 
-import { ButtonLink } from "../Buttons"
 import {
   StakingGlyphCentralizedIcon,
   StakingGlyphCloudIcon,
@@ -16,6 +15,7 @@ import {
   StakingGlyphTokenWalletIcon,
 } from "../icons/staking"
 import Translation from "../Translation"
+import { ButtonLink } from "../ui/buttons/Button"
 import { Center, Flex, VStack } from "../ui/flex"
 
 type SectionGridProps = ChildOnlyProp
@@ -134,7 +134,7 @@ const StakingHierarchy = () => {
                   eventName: "clicked solo staking",
                 })
               }}
-              width={{ base: "100%", md: "auto" }}
+              className="max-md:w-full"
             >
               {t("page-staking-more-on-solo")}
             </ButtonLink>
@@ -170,7 +170,7 @@ const StakingHierarchy = () => {
                   eventName: "clicked staking as a service",
                 })
               }}
-              width={{ base: "100%", md: "auto" }}
+              className="max-md:w-full"
             >
               {t("page-staking-more-on-saas")}
             </ButtonLink>
@@ -218,7 +218,7 @@ const StakingHierarchy = () => {
                   eventName: "clicked pooled staking",
                 })
               }}
-              width={{ base: "100%", md: "auto" }}
+              className="max-md:w-full"
             >
               {t("page-staking-more-on-pools")}
             </ButtonLink>
