+++ src/layouts/Tutorial.tsx
@@ -5,7 +5,6 @@ import { Badge } from "@chakra-ui/react"
 import type { ChildOnlyProp } from "@/lib/types"
 import type { MdPageContent, TutorialFrontmatter } from "@/lib/interfaces"
 
-import { ButtonLink } from "@/components/Buttons"
 import CallToContribute from "@/components/CallToContribute"
 import Card from "@/components/Card"
 import Codeblock from "@/components/Codeblock"
@@ -24,6 +23,7 @@ import {
 import TableOfContents from "@/components/TableOfContents"
 import TooltipLink from "@/components/TooltipLink"
 import TutorialMetadata from "@/components/TutorialMetadata"
+import { ButtonLink } from "@/components/ui/buttons/Button"
 import { mdxTableComponents } from "@/components/ui/table"
 import YouTube from "@/components/YouTube"
